go/koding: rename ec2info package to ec2dynamicdata
